+++ plots/extract_frames_from_gif.sh
@@ -1,0 +1,3 @@
+# Usage `sh extract_frames_from_gif.sh <file> <destination_dir>`
+mkdir -p "${1%.*}"
+ffmpeg -i "$1" "${1%.*}/frame_%d.png"